+++ packages/atomic/src/components/common/atomic-modal/atomic-modal.tsx
@@ -1,4 +1,3 @@
-// TODO: move to common components
 import {
   Component,
   h,