+++ packages/headless/src/features/facets/facet-set/interfaces/options.ts
@@ -2,9 +2,5 @@ import {FacetRequest} from './request';
 
 export type FacetOptionalParameters = Pick<
   FacetRequest,
-  | 'delimitingCharacter'
-  | 'filterFacetCount'
-  | 'injectionDepth'
-  | 'numberOfValues'
-  | 'sortCriteria'
+  'filterFacetCount' | 'injectionDepth' | 'numberOfValues' | 'sortCriteria'
 >;