+++ packages/headless/src/features/facets/range-facets/generic/interfaces/range-facet.ts
@@ -1,4 +1,5 @@
 import {RegisterDateFacetActionCreatorPayload} from '../../date-facet-set/date-facet-actions';
+import {DateFacetSetState} from '../../date-facet-set/date-facet-set-state';
 import {
   DateFacetRequest,
   DateRangeRequest,
@@ -16,9 +17,11 @@ import {
   NumericFacetResponse,
 } from '../../numeric-facet-set/interfaces/response';
 import {RegisterNumericFacetActionCreatorPayload} from '../../numeric-facet-set/numeric-facet-actions';
+import {NumericFacetSetState} from '../../numeric-facet-set/numeric-facet-set-state';
 
 export type RangeFacetRequest = DateFacetRequest | NumericFacetRequest;
 export type RangeValueRequest = DateRangeRequest | NumericRangeRequest;
+export type RangeFacetSetState = DateFacetSetState | NumericFacetSetState;
 export type RangeFacetRegistrationOptions =
   | RegisterDateFacetActionCreatorPayload
   | RegisterNumericFacetActionCreatorPayload;