+++ packages/headless/src/features/query/query-insight-analytics-actions.ts
@@ -1,14 +1,16 @@
 import {
   AnalyticsType,
+  InsightAction,
   makeInsightAnalyticsAction,
 } from '../analytics/analytics-utils';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 
-export const logSearchboxSubmit = makeInsightAnalyticsAction(
-  'analytics/searchbox/submit',
-  AnalyticsType.Search,
-  (client, state) =>
-    client.logSearchboxSubmit(
-      getCaseContextAnalyticsMetadata(state.insightCaseContext)
-    )
-);
+export const logSearchboxSubmit = (): InsightAction =>
+  makeInsightAnalyticsAction(
+    'analytics/searchbox/submit',
+    AnalyticsType.Search,
+    (client, state) =>
+      client.logSearchboxSubmit(
+        getCaseContextAnalyticsMetadata(state.insightCaseContext)
+      )
+  );